backend/requirements.txt
asgiref==3.11.0
beautifulsoup4==4.14.3
cachetools==6.2.4
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dj-rest-auth==7.0.2
Django==4.2.25
django-allauth==65.14.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
googleapis-common-protos==1.72.0
httplib2==0.31.1
idna==3.11
oauthlib==3.3.1
pillow==11.3.0
proto-plus==1.27.0
protobuf==6.33.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyparsing==3.3.1
python-dotenv==1.2.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
soupsieve==2.8.1
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
